+++ back/src/main/java/com/seguridadlimite/springboot/backend/apirest/services/AprendizServiceImpl.java
@@ -119,7 +119,8 @@ public class AprendizServiceImpl {
       		entity.setTienefirma("N");
   	  	}
 		entity.setIntentos(entity.getIntentos() == null ? 0 : entity.getIntentos());
-		entity.setContadorcertificados(entity.getContadorcertificados() == null ? 0 : entity.getContadorcertificados());
+		entity.setContadorcertificados(
+				entity.getContadorcertificados() == null ? (byte) 0 : entity.getContadorcertificados());
 		return dao.save(entity);
 	}
